Add files via upload
We explore a model for predicting stock prices of individual companies in the
financial sector using the historical returns of the NYSE Financial Sector Index (XLF), which tracks the performance of financial companies in the U.S. stock market. By applying a linear regression model, we examine the relationship between stock returns and financial sector returns, considering lagged returns and volatility.